relationship

+++ test_19_endpoints/src/routers/task.js
@@ -1,9 +1,10 @@
 const express = require('express')
 let {Task, newTask} = require('../../models/task.js')
 const router = express.Router()
+const auth = require('../middleware/auth.js')
 
-router.post('/task', async ( req, res ) => {
-    const task = newTask(req.body)
+router.post('/task', auth, async ( req, res ) => {
+    const task = newTask( {...req.body, owner: req.user._id})
     try {
         await task.save()
         return res.status(201).send( task )
@@ -11,26 +12,24 @@ router.post('/task', async ( req, res ) => {
         return res.status(400).send( e )
     }
 })
-router.get('/task', async ( req, res ) => {
+router.get('/task', auth, async ( req, res ) => {
     try {
-        return res.send( await Task.find({}) )
+        await req.user.populate('tasks').execPopulate()
+        return res.send( req.user.tasks )
     } catch (e) {
         return res.status(400).send( e )
     }
 })
-router.get('/task/:id', async (req, res) => {
+router.get('/task/:id', auth, async (req, res) => {
     try {
         const _id = req.params.id
-        const task = await Task.findById(  _id  )  
-        if( !task ) {
-            return res.status(404).send( ) 
-        }
+        const task = await Task.findOne({_id, owner: req.user._id})
         return res.send( task )
     } catch (e) {
         return res.status(400).send( e )
     }
 })
-router.patch('/task/:id', async (req, res) => {
+router.patch('/task/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body)
     const allowedUpdates = ['description', 'completed']
     const isValidOperation = updates.every( (update) => allowedUpdates.includes(update))
@@ -40,10 +39,9 @@ router.patch('/task/:id', async (req, res) => {
     try {
         const _id = req.params.id
         const body = req.body
-        // const task = await Task.findByIdAndUpdate(_id, body, {new: true, runValidators: true}) 
-        const task = await Task.findById(_id)
+        const task = await Task.findOne({_id, owner: req.user._id })
         if( !task ) {
-            return res.status(404).send( ) 
+            return res.status(404).send()
         }
         updates.forEach((up) => task[up] = body[up] )
         await task.save()
@@ -53,10 +51,10 @@ router.patch('/task/:id', async (req, res) => {
         return res.status(400).send( e )
     }
 })
-router.delete('/task/:id', async (req, res) => {
+router.delete('/task/:id', auth, async (req, res) => {
     try {
         const _id = req.params.id
-        const task = await Task.findByIdAndDelete(  _id  )  
+        const task = await Task.findByIdAndDelete({_id, owner: req.user._id }) 
         if( !task ) {
             return res.status(404).send( ) 
         }
